3_openai/NEW/1_basics_azure/pyproject.toml
[project]
name = "1-basics-azure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "openai>=1.66.3",
    "python-dotenv>=1.0.1",
]
